Move function to network_helpers
